+++ src/Proton/Codec/Encoders/Messaging/AmqpValueTypeEncoder.cs
@@ -36,7 +36,7 @@ namespace Apache.Qpid.Proton.Codec.Encoders.Messaging
          // Write the Array Type encoding code, we don't optimize here.
          buffer.WriteUnsignedByte((byte)EncodingCodes.Array32);
 
-         int startIndex = buffer.WriteOffset;
+         long startIndex = buffer.WriteOffset;
 
          // Reserve space for the size and write the count of list elements.
          buffer.WriteInt(0);
@@ -45,7 +45,7 @@ namespace Apache.Qpid.Proton.Codec.Encoders.Messaging
          WriteRawArray(buffer, state, values);
 
          // Move back and write the size
-         int endIndex = buffer.WriteOffset;
+         long endIndex = buffer.WriteOffset;
          long writeSize = endIndex - startIndex - sizeof(int);
 
          if (writeSize > Int32.MaxValue)
